Align startup script changes from version v0.4.0
From version `v0.4.0` on, this chart creates itself a config map `{{ template "localstack.fullname" . }}-init-scripts-config` which previous set ups. The script is now provided via `startupScriptContent`.

This PR documents this `startupScriptContent` field in the README, as it took me some time to understand the changes. Hope it helps others as well.