Allow setting the ingress class via `spec.ingressClassName`
Fixes: https://github.com/localstack/helm-charts/issues/114

The annotation `kubernetes.io/ingress.class` is deprecated and shouldn't
be used for specifying the ingress class.

Instead the option `ingressClassName` should be used. This introduces
that as an option to values and adds additional annotation comments for
enabling SSL passthrough when nginx is used as an ingress controller

As a result of this change, in local testing the following side effects
were observed: